+++ web/js/gerbera-items.module.js
@@ -108,6 +108,7 @@ export class BrowseItemView extends ItemView {
   constructor(pId) {
     super(pId, 'browse');
     $('#datagrid').show();
+    $('#database').hide();
   }
   get requestData() {
     let result = super.requestData;
@@ -126,6 +127,7 @@ export class SearchItemView extends ItemView {
     this.searchGerberaItems = this.searchGerberaItems.bind(this);
     $('#search').show();
     $('#datagrid').hide();
+    $('#database').hide();
     $('#searchButton').off('click').on('click', this.searchGerberaItems);
 
     $('#searchCaps').text(GerberaApp.serverConfig.searchCaps.split(',').join(', '));
@@ -166,6 +168,7 @@ export class FileItemView extends ItemView {
   constructor(pId) {
     super(pId, 'files');
     $('#datagrid').show();
+    $('#database').hide();
   }
   get requestData() {
     let result = super.requestData;
@@ -189,6 +192,7 @@ export class FileItemView extends ItemView {
 
 const initialize = () => {
   $('#datagrid').html('');
+  $('#database').hide();
   $('#search').hide();
   return Promise.resolve();
 };
@@ -228,11 +232,23 @@ const loadItems = (response, item) => {
   if (response.success) {
     const dataItems = currentItemView.dataItems(response);
     const datagrid = $('#datagrid');
+    const database = $('#database');
 
     if (datagrid.hasClass('grb-dataitems')) {
       datagrid.dataitems('destroy');
     }
-    datagrid.dataitems(dataItems);
+    if (!response.items || response.items.parent_id > 0) {
+      database.hide();
+      datagrid.show();
+      datagrid.dataitems(dataItems);
+    } else {
+      database.show();
+      const exportBtn = $('#exportButton');
+      const importBtn = $('#importButton');
+      exportBtn.off('click').on('click', { action: 'export', id: response.parent_id }, runExportAction);
+      importBtn.off('click').on('click', { action: 'import' }, runImportAction);
+      datagrid.hide();
+    }
     currentItemView.totalMatches = dataItems.parentItem.total_matches;
     Trail.makeTrailFromItem(dataItems.parentItem, item);
   }
@@ -240,6 +256,47 @@ const loadItems = (response, item) => {
 
 const setPage = (pageNumber) => {
   GerberaApp.setCurrentPage(pageNumber);
+};
+
+const runExportAction = (event) => {
+  var requestData = {
+    req_type: 'action',
+    action: event.data.action
+  };
+  $('#exportMissing').hide();
+  $('#exportCode').hide();
+  requestData[Auth.SID] = Auth.getSessionId();
+  $.ajax({
+    url: GerberaApp.clientConfig.api,
+    type: 'get',
+    data: requestData
+  })
+    .then((response) => {
+      if (response.success && response.items && response.items.item) {
+        const exportCode = $('#exportCode');
+        exportCode.text(JSON.stringify(response.items.item, null, 4));
+        exportCode.show();
+      } else {
+        $('#exportMissing').show();
+      }
+    })
+    .catch((err) => GerberaApp.error(err));
+};
+
+const runImportAction = (event) => {
+  var requestData = {
+    req_type: 'action',
+    action: event.data.action,
+  };
+  requestData[Auth.SID] = Auth.getSessionId();
+  requestData['file_name'] = $('#importFile').val();
+  $.ajax({
+    url: GerberaApp.clientConfig.api,
+    type: 'get',
+    data: requestData
+  })
+    .then((response) => { console.log(response); })
+    .catch((err) => GerberaApp.error(err));
 };
 
 const nextPage = (event) => {